fix: relocate taxonomy tests to satisfy pkg-file-count gate
